Add QDMA support for AU280
This commit adds support for the QDMA core of Xilinx as an alternative to our custom BlueDMA core for the Alveo U280 card.

This requires some changes in the hardware design itself. The QDMA is built around the PCIe subsystem directly and remains in our host subsystem. We use it in "descriptor bypass" mode, and generate the descriptor for DMA transfers directly in user logic with our QDMADescriptorGenerator.
We also need a slightly different InterruptController since the QDMA uses another interrupt interface for user interrupts. This QDMAIntrCtrl only handles user interrupts since DMA interrupts are directly sent by the QDMA interrupt core itself.
Currently this design only supports 28 user interrupts since the QDMA only supports 32 interrupts, of which 4 are reserved for DMA interrupts. We are currently investigating whether we can extend the QDMA core or will implement a solution with shared interrupts similar to the AWS interrupt controller.

The user space runtime is not changed since the descriptor generator offers the same control register layout as BlueDMA.

The driver is mainly extended with means to initialize and shutdown the QDMA core during loading an unloading the driver module. Also DMA interrupts are acknowledged slightly different.

In the BlueDMA PCIe design, the second BAR is activated as well to have a matching BAR configuration between both variants